% Auto-generated: do not edit by hand
\name{dashFeatureViewer}

\alias{dashFeatureViewer}

\title{DashFeatureViewer component}

\description{
Feature Viewer for Protein or DNA sequences From Calipho  Add Features as object Selected Feature or Sequence available in callback
}

\usage{
dashFeatureViewer(id=NULL, options=NULL, sequence=NULL, features=NULL,
viewerStyle=NULL, zoom=NULL, selectedSeq=NULL, darkMode=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{options}{Named list. Options for Feature Viewer.}

\item{sequence}{Character. The Sequence or integer length value.}

\item{features}{Unnamed list. The Features to View.}

\item{viewerStyle}{Named list. The Style of Viewer.}

\item{zoom}{Unnamed list. The Zoom of Viewer.}

\item{selectedSeq}{Unnamed list. The selected sequence of Viewer.}

\item{darkMode}{Logical. The selected sequence of Viewer.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

